Add ttyin/ttyout as stdio streams.
These streams can be used when stdin/stdout are piped, e.g. when using
SSH to create a secure stream to a remote server. In that case
ttyin/ttyout can still be wired to the terminal input and output. In the
case of ssh, this can be used to print the password prompt, and read the
password from the user.

In the normal case, these streams will be unused.