Make builtinsFS less special
Give builtinsFS a distinct name (used to be a duplicate 'jsfs') and move it under the main FileSystemManager, but hide it from the end user (e.g. when running `mount`). This eliminates the need for special functions to find the builtins, which also fixes the introduced problem in the redirection PR (#226), where any command entered on the command line is looked up in a universal way inside ExecuteContext.